Removed Bugs Caused Restructuring
