+++ src/App.js
@@ -1,26 +1,14 @@
 import React, { useState } from 'react';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import CollegeAdmin from './CollegeAdmin';
-import Company from './Company';
-import { makeStyles, useTheme } from '@material-ui/core/styles';
+import CollegeAdmin from './CollegeAdmin/CollegeAdmin';
+import Company from './Company/Company';
 import { BrowserRouter as Router } from 'react-router-dom';
-// import Temp from './temp'
-/**
- * Template File
- */
+
 function App() {
-	const drawerWidth = 240;
-	const useStyles = makeStyles((theme) => ({
-		root: {
-			display: 'flex',
-		},
-	}));
+	const [userRole, setUserRole] = useState('company');
 
-	const [userRole, setUserRole] = useState('admin');
-
-	const classes = useStyles();
 	return (
-		<div className={classes.root}>
+		<div>
 			<CssBaseline /> {/*This is to normailze the CSS (Builtin component)*/}
 			<Router>{userRole == 'admin' ? <CollegeAdmin /> : <Company />}</Router>
 		</div>
