+++ src/Company/Company.js
@@ -1,14 +1,20 @@
 import React from 'react';
 import MainContent from './components/MainContent';
 import Header from './components/Header';
+import { makeStyles } from '@material-ui/core/styles';
 
 function Company() {
+	const useStyles = makeStyles((theme) => ({
+		main: {
+			display: 'flex',
+		},
+	}));
+	const classes = useStyles();
 	return (
-		<div>
+		<div className={classes.main}>
 			<Header />
 			<MainContent />
 		</div>
 	);
 }
-
 export default Company;
