+++ src/Company/containers/index.js
@@ -1,8 +1,8 @@
 export { default as CreateNotice } from './CreateNotice';
 export { default as CurrentDatails } from './CurrentDatails';
-export { default as cHome } from './Home';
-export { default as cQnA } from './QnA';
+export { default as Home } from './Home';
+export { default as QnA } from './QnA';
 export { default as RequestDate } from './RequestDate';
 export { default as StartPlacement } from './StartPlacement';
 export { default as ViewNotice } from './ViewNotice';
-export { default as cViewPlacements } from './ViewPlacements';
+export { default as ViewPlacements } from './ViewPlacements';
